git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/diffCond-prop@38915 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/object/SMGRegion.java
@@ -21,7 +21,7 @@ public final class SMGRegion extends SMGObject implements SMGObjectTemplate {
     super(pSize, pLabel, SMGObjectKind.REG);
   }
 
-  public SMGRegion(int pSize) {
+  public SMGRegion(long pSize) {
     super(pSize, "ID" + SMGCPA.getNewValue(), SMGObjectKind.REG);
   }
